requirements.txt
flask
openai-whisper
torch
ffmpeg-pythonpython-dotenv